pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pcdl"
version = "1.0.0"
description = "Persistence Scene Change Detection for PSCDL 2026"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "PSCDL Team"},
]
dependencies = [
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "Pillow>=10.0.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pydensecrf>=2.2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["pcdl*"]
